chore: Update changelog generation to include zigbee2mqtt-windfront
