+++ package.json
@@ -66,7 +66,7 @@
         "zigbee-herdsman": "6.0.1",
         "zigbee-herdsman-converters": "25.6.0",
         "zigbee2mqtt-frontend": "0.9.20",
-        "zigbee2mqtt-windfront": "^1.8.1"
+        "zigbee2mqtt-windfront": "1.8.1"
     },
     "devDependencies": {
         "@biomejs/biome": "^2.1.4",
